% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\name{im_cols}
\alias{im_cols}
\title{Function to extract colors as hex codes}
\usage{
im_cols(...)
}
\arguments{
\item{...}{Character names of colors}
}
\value{
A named character vector of hex codes
}
\description{
Function to extract colors as hex codes
}
\examples{
im_cols("RICH BROWN")
im_cols("EARTH GREEN", "PALE BROWN")

}
